Corrected calculation of accuracy and f1 score during validation when using binary output Pytorch-lighting requires float tensors for binary classification

+++ mef/utils/pytorch/lighting/module.py
@@ -96,8 +96,8 @@ class MefModel(pl.LightningModule):
         else:
             y = torch.argmax(y, dim=-1)
 
-        self._accuracy(output, y)
-        self._f1_macro(output, y)
+        self._accuracy(output.float(), y)
+        self._f1_macro(output.float(), y)
         return
 
     def validation_step(self,
